chats transferred to redux, chat list container added, works good
