+++ JulianaJarm/package.json
@@ -31,6 +31,7 @@
     "prop-types": "^15.7.2",
     "proptypes": "^1.1.0",
     "react": "^16.13.1",
+    "react-addons-update": "^15.6.2",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
     "react-scroll": "^1.7.16",
